torchrun for DDP

+++ xuance/torch/agents/base/agent.py
@@ -242,6 +242,7 @@ class Agent(ABC):
         if self.distributed_training:
             if dist.get_rank() == 0:
                 if os.path.exists(self.learner.snapshot_path):
-                    os.remove(os.path.join(self.learner.snapshot_path, "snapshot.pt"))
+                    if os.path.exists(os.path.join(self.learner.snapshot_path, "snapshot.pt")):
+                        os.remove(os.path.join(self.learner.snapshot_path, "snapshot.pt"))
                     os.removedirs(self.learner.snapshot_path)
             destroy_process_group()
